raspberry-pi-code/requirements.txt
requests==2.28.2
pyserial==3.5
